Fix perms checks and Launch init to work with Win 7
Windows 7 doesn't like the SDDI format of the bluetooth permissions
string we need in the registry, and will also throw if it hits any
code it can't deal with, like Win10 UWP stuff we use for bluetooth.
Wrap these in version guards that will allow them to be skipped on
anything < Win 10 15063.