fix: Correct client/device mismatching
I took a few attampts at this, but eventually worked out that
connects were blotting eachother and not resetiing the device
list, so the devices were referring to the wrong clients.

This resolved the #168 like crashes I could reproduce.

I also abstracted out the error recording, since there a lot of
places that exceptions could be thrown and should just be
handled by logging.

The disconnect system also got tweaked, so that all
disconnects are handled in the same way.