Experimental Feature: This commit adds an experimental feature that allows a gradual ramping up of the intensity of the FunStim device output over the duration of a video.
A setting of 0% implies to ramp up. A setting of 100% implies a ramp up
from zero intensity to full intensity over the duration of the video.

This is my first try to add a major feature, and it might be incomplete
or not the correct way to implement.

Definitely needs testing! Comments Welcome.